Give up on using the a user systemd unit
We encountered an issue where invoking docker start within the
ExecStart command of a user systemd unit did not have permission to
open the docker sock. It's not obvious what is causing this, but it's
not worth the effort to debug because the docker noncontainer lifecycle
commands work fine for inspecting and controlling the container.