Fix the dockerfile caching
Previously, all of the ARGs were at the top, which caused the cache of
the early RUN commands to be invalidated when we set the REPO_CACHE_ID.

This relocates the ARGs to close to their first uses.

We also add a `git pull` to ensure that the specified branch is
up-to-date.