Upgrade vanilla resources
Previously, we were avoiding clobbering any config file that previously
existed. But that limits the ability to pick up changes automatically
upon re-running the playbook. This change, instead, will automatically
upgrade any config file that had been unchanged since it was installed
by a previous run of the playbook.

A known issue with this implementation is that it is sensitive to
whitespace changes in the config file.