Use the refresh-conda-env script during bootstrapping
This change flows from the need to specify a Python version that matches
the default Python version of the Ubuntu LTS release. We want to match
that because that's the version that our ROS version expects.

For readability, we've migrated the jugglebot Conda environment setup
into an Ansible playbook. That change necessitated building a bootstrap
Conda environment to run that playbook.

Along the way, we moved the yq installation into a separate tasks file
and move the version lookup dicts into vars file (as opposed to being
buried in a Jinja2 template file).