Add the `conda depgraph` command
This `conda depgraph` tool is pretty helpful for understanding what
caused a package to be installed. Unfortunately, it depends on Java,
which is a large download. We're adding the environment config file. But
we won't create the environment during host provisioning. We can create
the environmnent when we encounter the need for the tool.