Rename refresh script, add fnm and set the Node.js version
The refresh-conda-env script has expanded in scope to include Node.js
dependencies, so we've renamed it.

Along the way, I decided to get rid of the magic creation of a file
that's ignored by git: namely conda_env.yml. That file is now generated
into the ~/.jugglebot/host_setup/defaults directory so that only the
source of truth is readily visible in the repo.

This doesn't yet support bash completions and prompt indicators for
Node.js because I want to ensure that they won't be too slow.