Use similar refactored config management
Previously, we refactored the WSL provisioning to construct an overlayed
config object so that we can automatically bring in newly introduced
default config values. This introduces the same config construction
scheme within the docker container builds.

Note that the docker container builds inherit the config from the parent
WSL environment except that we override the ROS version codename such
that it will match the version of ROS that has tier 1 support for the
specified Ubuntu release of the base image.